fix missing token updates
PushToken values can change and this wasn't being account for in
checkAndTriggerSubscriptionChanged(). This fixes a bug where the backend
would not get the new token when toggling notification push off and on
then refreshing the page.

Also logging instead of throwing when there is no push sub yet when
updatePushSubscriptionNotificationTypes is called. This isn't an error
as the value is simply populated when the sub is created.