fix confirm deliveries & notification open
Both confirm deliveries & notification open were building their REST API
with an old placeholder deviceId value from Database.getSubscription().
To fix the problem we are now using the model cache where we started to
store the subscription id with v16.

This commit is scoped to fixing this deviceId issue with confirm
deliveries & notification open only. A larger change is needed to
clean this up completely.