fix keyPath issue with a new table & migrate data
Bumping indexDB to v5 and added a migration.

This commit address the two states NotificationClicked's keyPath can be
in due an older migration being changed when it shouldn't have. See
the previous commit for more details.

We are doing the following indexedDB table renames in this commit:
* NotificationClicked -> Outcomes.NotificationClicked
* NotificationReceived -> Outcomes.NotificationReceived

IndexedDB does not allow changing the keyPath to correct it so instead
we must create a new table. Only "NotificationClicked" needs to be fixed
however we are also renaming "NotificationReceived " and prefixing them
with "Outcomes." so they are consistent.

Comments in code explain in detail the effected versions and what was
wrong with them.