rm soft-prompt in Notifications.requestPermission
Developers want to directly display the native notification permission
prompt. However we were trying to detect if a user interaction was
required without considering if the developer already met it.

There isn't a browser API to detect if the user interaction was met
at either the global level or stacktrace to detect this for them. There
also isn't any errors thrown or returned to show one as a fallback either.
We don't want to require a parameter so instead simply just always attempt
to show the native prompt.