add example of a root service worker
This is to test OneSignal with conflicting a root service worker. This
causes the OneSignal SDK to sometimes not include a push token
when it makes a REST API to POST /users. This is due to the OneSignal
SDK sometimes getting the registration for the root service worker due
to a race condition, instead of the OneSignal service worker.

We will fix this problem in a follow up commit.