fix race condition getting wrong sw registration
If you attempt to get the service worker registration immediately after
registering it the browser can give you a different registration than
exepected. This is due to the fact the new registeration may not be
ready yet and there migth be sw in scope at a higher level.

The problem noted above cause the OneSignal SDK to register for a push
token on the wrong service worker. This in turn causes the SDK to
attempt to create a user with a push susbcription but without a push
token, which results in a 400 error on POST /users.

To fix, instead of calling navigator.serviceWorker.getRegistration
we are now simply using the ServiceWorkerRegistration give to us by
navigator.serviceWorker.register to avoid this race condition completely