Clear operation queue for operation repo tests
Running tests consecutively that enqueue operations to the SubscriptionExecutor will fail due to the operationQueue length not matching the expected value. pushSubscription, smsSubscription, emailSubscription now point to the same SubscriptionExecutor and the operationQueue was not being cleared for the next test. The other tests don't fail but I also cleared the operation queue there too.