Add new records state
Used to keep track of creation of users and subscriptions inorder to delay operations afterwards. The delay is needed because after the server creates the user/subscription, there is a small amount of time where the data is not up to date. Ex: fetching a user right after a user create may return an outdated external id.