Use local external Id in user hydration
The server may still be updating and can return an outdated external Id from the getUser request. Instead of waiting for the updated external Id, we can hydrate the user with the local external Id as getting a response from the server means that the login was successful.

The external Id param from login() is used as the local external Id instead of the Identity Model because the Identity Model gets reset from resetModelRepoAndCache() on L89 and returns undefined.