Prevent race with switching users & handleOnFocus
If the site focus changes when the login or logout fn is still running
we can get errors with missing ids. To prevent this always await to
ensure they are completed.

The most common way this can happen is if you call window.prompt()
or window.alert() with 100ms before or after calling OneSignal.login()