Add an api call for web config if no default url
There was an issue with legacy HTTP integrations where the
defaultNotificationUrl was never saved in the DB. To account for this,
we try to get the default URL from the app config on notification click
and save it to the DB for the future use.

Choosing between notification received and clicked to add this logic,
decided on notification clicked to avoid doing extra api call when the
user may never click the notification.