test(IDManager): update uuid sample to be more similar to a realist shape
The previous sample used a simple string, e.g. u1, whereas the updated version uses a 5-part string delimited by hyphens. This is closer to the representation of a UUID, which has a "a 128-bit value commonly represented as a string of 32 hexadecimal digits, separated into five groups by hyphens in the format (xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx)"